Added support for specifying whether to use Growl or NSUserNotificationCenter.
This will be useful if you want to provide an option to your users to use Growl or NSUserNotificationCenter rather than using NSUserNotificationCenter over Growl. SCNotificationCenter will save the chosen notification method to your application's preferences file. The default behaviour is to use NSUserNotificationCenter if it's available and if not, use Growl.